pyproject.toml
[project]
name = "adk-agent-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.18.0",
]
